+++ api-functions/ingredients_upload.php
@@ -36,7 +36,6 @@ foreach ($data['ingredients'] as $ingredient) {
     $notes = mysqli_real_escape_string($conn, $ingredient['notes'] ?? null);
     $profile = mysqli_real_escape_string($conn, $ingredient['profile'] ?? null);
     $solvent = mysqli_real_escape_string($conn, $ingredient['solvent'] ?? null);
-    $odor = mysqli_real_escape_string($conn, $ingredient['odor'] ?? null);
     $allergen = (int)($ingredient['allergen'] ?? null);
     $flavor_use = (int)($ingredient['flavor_use'] ?? null);
     $soluble = mysqli_real_escape_string($conn, $ingredient['soluble'] ?? null);
@@ -82,14 +81,14 @@ foreach ($data['ingredients'] as $ingredient) {
     // Create SQL query for inserting data
     $query = "INSERT INTO ingredients 
         (name, INCI, type, strength, category, purity, cas, einecs, FEMA, reach, tenacity, chemical_name, 
-        formula, flash_point, appearance, rdi, notes, profile, solvent, odor, allergen, flavor_use, soluble, 
+        formula, flash_point, appearance, rdi, notes, profile, solvent, allergen, flavor_use, soluble, 
         logp, cat1, cat2, cat3, cat4, cat5A, cat5B, cat5C, cat5D, cat6, cat7A, cat7B, cat8, cat9, cat10A, 
         cat10B, cat11A, cat11B, cat12, impact_top, impact_heart, impact_base, created_at, updated_at, owner_id, 
         usage_type, noUsageLimit, byPassIFRA, isPrivate, molecularWeight, physical_state, cid, shelf_life, aromaTrackID)
         VALUES 
         ('$name', '$INCI', '$type', '$strength', $category, '$purity', '$cas', '$einecs', '$FEMA', '$reach', 
         '$tenacity', '$chemical_name', '$formula', '$flash_point', '$appearance', $rdi, '$notes', '$profile', 
-        '$solvent', '$odor', $allergen, $flavor_use, '$soluble', '$logp', $cat1, $cat2, $cat3, $cat4, $cat5A, 
+        '$solvent', $allergen, $flavor_use, '$soluble', '$logp', $cat1, $cat2, $cat3, $cat4, $cat5A, 
         $cat5B, $cat5C, $cat5D, $cat6, $cat7A, $cat7B, $cat8, $cat9, $cat10A, $cat10B, $cat11A, $cat11B, 
         $cat12, '$impact_top', '$impact_heart', '$impact_base', NOW(), NOW(), '$userID', '$usage_type', 
         $noUsageLimit, $byPassIFRA, $isPrivate, '$molecularWeight', $physical_state, $cid, $shelf_life, '$aromaTrackID')";